+++ packages/fether-react/src/App/App.js
@@ -20,7 +20,6 @@ import { version } from '../../package.json';
 import Accounts from '../Accounts';
 import BackupAccount from '../BackupAccount';
 import Onboarding from '../Onboarding';
-import RequireParityVersion from '../RequireParityVersion';
 import BackupPhrase from '../BackupPhrase';
 import RequireHealthOverlay from '../RequireHealthOverlay';
 import Send from '../Send';
@@ -157,8 +156,12 @@ class App extends Component {
                   component={Whitelist}
                 />
                 <Route
-                  path='/backup/:accountAddress'
+                  path='/backup/:accountAddress/'
                   component={BackupAccount}
+                />
+                <Route
+                  path='/backupPhrase/:accountAddress/:showWarning'
+                  component={BackupPhrase}
                 />
                 <Route
                   path='/send/:tokenAddress/from/:accountAddress'
